+++ src/utils/test_data.c
@@ -358,7 +358,7 @@ void select_node(Cursor* cursor) {
     };
     Node node = {0, 2, "l", "person"};
     for (int i = 0; i < 4; i++) {
-        strlcpy(node.type, data[i].type, NAME_TYPE_WITH_TERM_LENGTH);
+        strlcpy(node.type, data[i].type, (size_t)NAME_TYPE_WITH_TERM_LENGTH);
         Iterator* iterator = select_all_nodes(cursor, &(node));
         while (has_next(iterator)) {
             print_node(next(iterator));
@@ -375,7 +375,7 @@ void select_relationship(Cursor* cursor) {
     };
     Relationship rel = {0, 2, 0, "", "", ""};
     for (int i = 0; i < 4; i++) {
-        strlcpy(rel.type, data[i].type, NAME_TYPE_WITH_TERM_LENGTH);
+        strlcpy(rel.type, data[i].type, (size_t)NAME_TYPE_WITH_TERM_LENGTH);
         fail_print("Find %s", rel.child_type);
         Iterator* iterator = select_all_relationships(cursor, &(rel));
         while (has_next(iterator)) {
@@ -391,7 +391,7 @@ void select_property(Cursor* cursor) {
     };
     Property pr = {0, 2, INT, "", "", ""};
     for (int i = 0; i < 2; i++) {
-        strlcpy(pr.type, data[i].type, NAME_TYPE_WITH_TERM_LENGTH);
+        strlcpy(pr.type, data[i].type, (size_t)NAME_TYPE_WITH_TERM_LENGTH);
         Iterator* iterator = select_all_properties(cursor, &(pr));
         while (has_next(iterator)) {
             print_property(next(iterator));
@@ -514,8 +514,8 @@ void create_node_smoke(Cursor* cursor, int num) {
     for (int j = 0; j < 500 * num; j++) {
         int r = rand() % 15;
 
-        strlcpy(name, "sasaovchsa", 10);
-        strlcpy(node->type, types[r], NAME_TYPE_WITH_TERM_LENGTH);
+        strlcpy(name, "sasaovchsa", (size_t)10);
+        strlcpy(node->type, types[r], (size_t)NAME_TYPE_WITH_TERM_LENGTH);
 
         node->id = 0;
         node->name_length = 11;
@@ -571,8 +571,8 @@ void delete_smoke_test(Cursor* cursor) {
         for (int j = 0; j < 1 * i; j++) {
             int r = rand() % 15;
 
-            strlcpy(name, "sasaovchsa", 10);
-            strlcpy(node->type, types[r], NAME_TYPE_WITH_TERM_LENGTH);
+            strlcpy(name, "sasaovchsa", (size_t)10);
+            strlcpy(node->type, types[r], (size_t)NAME_TYPE_WITH_TERM_LENGTH);
 
             node->id = 0;
             node->name_length = 11;
@@ -623,8 +623,8 @@ void update_smoke_test(Cursor* cursor) {
         for (int j = 0; j < 100; j++) {
             int r = rand() % 15;
 
-            strlcpy(name, "sasaovchsa", 10);
-            strlcpy(node->type, types[r], NAME_TYPE_WITH_TERM_LENGTH);
+            strlcpy(name, "sasaovchsa", (size_t)10);
+            strlcpy(node->type, types[r], (size_t)NAME_TYPE_WITH_TERM_LENGTH);
 
             node->id = 0;
             node->name_length = 11;
@@ -652,8 +652,8 @@ void select_delete_smoke_test(Cursor* cursor) {
     Node *node = (Node*) malloc(NODE_SIZE);
     char* name = malloc(11 * CHAR_SIZE);
 
-    strlcpy(name, "sasaovch", NAME_TYPE_WITH_TERM_LENGTH);
-    strlcpy(node->type, "person", NAME_TYPE_WITH_TERM_LENGTH);
+    strlcpy(name, "sasaovch", (size_t)NAME_TYPE_WITH_TERM_LENGTH);
+    strlcpy(node->type, "person", (size_t)NAME_TYPE_WITH_TERM_LENGTH);
 
     node->id = 0;
     node->name_length = 11;
@@ -697,8 +697,8 @@ void create_n_nodes_(Cursor* cursor, int n) {
     Node *node = (Node*) malloc(NODE_SIZE);
     char* name = malloc(11 * CHAR_SIZE);
 
-    strlcpy(name, "sasaovch", NAME_TYPE_WITH_TERM_LENGTH);
-    strlcpy(node->type, "person", NAME_TYPE_WITH_TERM_LENGTH);
+    strlcpy(name, "sasaovch", (size_t)NAME_TYPE_WITH_TERM_LENGTH);
+    strlcpy(node->type, "person", (size_t)NAME_TYPE_WITH_TERM_LENGTH);
 
     node->id = 0;
     node->name_length = 11;
